@stk_model_gpposterior: Preserve row and column names
* core/@stk_model_gpposterior/stk_model_gpposterior.m: Preserver
  row and column names in input/output data.
* core/@stk_model_gpposterior/stk_model_update.m: Idem.
* core/@stk_model_gpposterior/stk_predict.m: Use double() to
  extract the numerical content of the 'output_data' field.

+++ core/@stk_model_gpposterior/stk_model_update.m
@@ -28,8 +28,8 @@
 
 function M = stk_model_update (M, xi, zi)
 
-M.input_data = [M.input_data; double(xi)];
-M.output_data = [M.output_data; double(zi)];
+M.input_data = [M.input_data; xi];
+M.output_data = [M.output_data; zi];
 
 % FIXME: use @stk_kreq/stk_update ?
 M.kreq = stk_kreq_qr (M.prior_model, M.input_data);
